[Bronze V] Title: 이상한 기호, Time: 224 ms, Memory: 17672 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,12 +4,12 @@ public class Main {
     
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
-        int a = in.nextInt();
-        int b = in.nextInt();
+        long a = in.nextLong();
+        long b = in.nextLong();
         
         in.close();
         
-        int answer = (a + b) * (a - b);
+        long answer = (a + b) * (a - b);
         System.out.print(answer);
     }
 }
